Delete all namespaces on resource service on DeleteAllUserNamespaces

+++ pkg/server/namespace.go
@@ -484,6 +484,10 @@ func (s *Server) DeleteAllUserNamespaces(ctx context.Context) error {
 			return unsubErr
 		}
 
+		if delErr := s.clients.Resource.DeleteAllUserNamespaces(ctx); delErr != nil {
+			return delErr
+		}
+
 		// kube-api don`t have method to delete list of namespaces
 		for _, ns := range deletedNamespaces {
 			if delErr := s.clients.Kube.DeleteNamespace(ctx, kubeAPIModel.NamespaceWithOwner{Name: ns.ID}); delErr != nil {
